feat(filter-lab): move the lab into the filters panel
滤镜实验室 browses and applies filters, so it belongs beside the filter
library, not under adjustments. The lab plus its adjustment-layer apply
wiring moves as one shelf component; the filters panel hosts it as a
third mode and adjustments keeps only 我的/LUT.
